Move generic components to common directory
