+++ components/common/index.js
@@ -1,5 +1,10 @@
 import Container from './Container';
+import { PrimaryButton, SecondaryButton } from './Button';
+import { Input } from './Form';
 
 export {
   Container,
+  PrimaryButton,
+  SecondaryButton,
+  Input,
 };
